chore(snippets): update generated types and methods

+++ src/resources/snippets/snippets.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../resource';
+import { isRequestOptions } from '../../core';
 import * as Core from '../../core';
 import * as ContentAPI from './content';
 import { Content, ContentGetParams } from './content';
@@ -28,7 +29,7 @@ export class Snippets extends APIResource {
     params: SnippetUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<SnippetUpdateResponse> {
-    const { zone_id, ...body } = params;
+    const { zone_id = this._client.zoneId, ...body } = params;
     return (
       this._client.put(
         `/zones/${zone_id}/snippets/${snippetName}`,
@@ -41,10 +42,20 @@ export class Snippets extends APIResource {
    * Fetches all snippets belonging to the zone.
    */
   list(
-    params: SnippetListParams,
+    params?: SnippetListParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<SnippetListResponsesV4PagePaginationArray, SnippetListResponse>;
+  list(
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<SnippetListResponsesV4PagePaginationArray, SnippetListResponse>;
+  list(
+    params: SnippetListParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<SnippetListResponsesV4PagePaginationArray, SnippetListResponse> {
-    const { zone_id, ...query } = params;
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { zone_id = this._client.zoneId, ...query } = params;
     return this._client.getAPIList(`/zones/${zone_id}/snippets`, SnippetListResponsesV4PagePaginationArray, {
       query,
       ...options,
@@ -56,10 +67,19 @@ export class Snippets extends APIResource {
    */
   delete(
     snippetName: string,
-    params: SnippetDeleteParams,
+    params?: SnippetDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<SnippetDeleteResponse | null>;
+  delete(snippetName: string, options?: Core.RequestOptions): Core.APIPromise<SnippetDeleteResponse | null>;
+  delete(
+    snippetName: string,
+    params: SnippetDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<SnippetDeleteResponse | null> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(snippetName, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return (
       this._client.delete(`/zones/${zone_id}/snippets/${snippetName}`, options) as Core.APIPromise<{
         result: SnippetDeleteResponse | null;
@@ -72,10 +92,19 @@ export class Snippets extends APIResource {
    */
   get(
     snippetName: string,
-    params: SnippetGetParams,
+    params?: SnippetGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<SnippetGetResponse>;
+  get(snippetName: string, options?: Core.RequestOptions): Core.APIPromise<SnippetGetResponse>;
+  get(
+    snippetName: string,
+    params: SnippetGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<SnippetGetResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(snippetName, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return (
       this._client.get(`/zones/${zone_id}/snippets/${snippetName}`, options) as Core.APIPromise<{
         result: SnippetGetResponse;
@@ -155,7 +184,7 @@ export interface SnippetUpdateParams {
   /**
    * Path param: Use this field to specify the unique ID of the zone.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param: Provide metadata about the snippet.
@@ -181,21 +210,21 @@ export interface SnippetListParams extends V4PagePaginationArrayParams {
   /**
    * Path param: Use this field to specify the unique ID of the zone.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface SnippetDeleteParams {
   /**
    * Use this field to specify the unique ID of the zone.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface SnippetGetParams {
   /**
    * Use this field to specify the unique ID of the zone.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 Snippets.SnippetListResponsesV4PagePaginationArray = SnippetListResponsesV4PagePaginationArray;
